# Faster campaigns tests, part 3: clean up & parallelize